formula

+++ repository/src/main/java/com/softserve/edu/device/test/data/DeviceTestData.java
@@ -41,7 +41,7 @@ public interface DeviceTestData {
     double getTestCorrectedCurrentConsumption(int testIndex);
     double getTestCumulativeImpulsesValueWithoutCorrection(int testIndex);
     double getTestCurrentConsumptionWithoutCorrection(int testIndex);
-    long getTestEstimatedError(int testIndex);
+    double getTestEstimatedError(int testIndex);
     double getTestInitialCounterValue(int testIndex);
     double getTestTerminalCounterValue(int testIndex);
     long getTestUnixBeginTime(int testIndex);
